refactor: ref createGroup
